LayoutTests/fast/reflections/reflection-overflow-scroll-expected.txt

Check if the size of reflected images are considered as overflow size. See this bug for detail: https://bugs.webkit.org/show_bug.cgi?id=27979.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.scrollLeft is 150
PASS element.scrollTop is 80
PASS element.scrollLeft is 150
PASS element.scrollTop is 80
PASS element.scrollLeft is 150
PASS element.scrollTop is 80
PASS successfullyParsed is true

TEST COMPLETE
